feat: adds support for simple env var lookup of creds if none found otherwise
Modifies the default keychain to use all existing mechanisms first, but
before giving up and returning an Anonymous user identity for further
auth flow (e.g. token issuance), it will do a final check for the
environment variables:

CRANE_USERNAME
CRANE_PASSWORD

If the username is found, those will be used as the identity to pass to
the auth flow (tokens etc).

Signed-off-by: Zach Hill <zach@anchore.com>